+++ src-tauri/src/lib.rs
@@ -395,8 +395,8 @@ pub fn run() {
             audio::commands::audio_set_volume,
             audio::commands::audio_update_replay_gain,
             audio::commands::audio_set_eq,
-            audio::commands::autoeq_entries,
-            audio::commands::autoeq_fetch_profile,
+            audio::autoeq_commands::autoeq_entries,
+            audio::autoeq_commands::autoeq_fetch_profile,
             audio::commands::audio_preload,
             audio::commands::audio_play_radio,
             audio::preview::audio_preview_play,
